+++ crates/core/tests/java/method_parameters_wrap/brace_overflow.java
@@ -1,0 +1,5 @@
+class Sample {
+    public void methodName(int alphaParameter, int betaParameter, int gammaParameter) {
+        int x = 1;
+    }
+}