+++ crates/core/tests/java/method_parameters_wrap/brace_overflow_next_line_brace.out.java
@@ -1,0 +1,7 @@
+class Sample {
+    public void methodName(int alphaParameter, int betaParameter, int gammaParameter)
+    {
+        int x = 1;
+    }
+}
+